general coding convention changes (related to #11).
due to the recent changes, there are many `for` loops containing a single `if`
statement which is only used to narrow the loop targets. due to the number of
such constructions, from now on we will write them in the single line as long
as two statements are strongly correlated.

--HG--
branch : 2.0