added default key mapping for Pop'n Music Controller 2.
this mapping is known to make the UI buggy (the game engine itself
does not); axis #1 (mapped to the button 8) does not return to
the origin when unpressing the button, but instead goes to
the opposite side of the axis and keeps doing so. maybe we will
need some way to customize the joystick's behavior.